Move interface definition.
